+++ docs/src/release_notes.rst
@@ -14,6 +14,8 @@ Unreleased
 - Fix attribute and call after walrus leading parser error (:issue:`174`)
 - Fix parsing error in doctest blocks with empty lines (:issue:`176`)
 - Improve error message on uncaught parsing errors (:issue:`177`)
+- Add ``level`` argument to :rst:dir:`autolink-preface` to support
+  file-level prefaces (:issue:`180`)
 
 0.17.0 (2025-02-18)
 -------------------